Refactor game logic and improve styling approach
The main refactoring focused on improving the game's logic - unnecessary variables were removed and the logic for checking game availability was significantly changed. The method for swapping fields was also updated to not throw an exception when there's no argument passed, instead, such cases are now handled differently. In addition to logic improvements, the import approach for SCSS in GameField.vue was also modified to provide a more compact and neat format.